Autorização Atualizado

+++ DesafioFinal.Server/Controllers/UsuarioController.cs
@@ -1,8 +1,10 @@
 ﻿using DesafioFinal.Server.Data;
 using DesafioFinal.Server.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Conventions;
 
 namespace DesafioFinal.Server.Controllers
 {
@@ -37,10 +39,18 @@ namespace DesafioFinal.Server.Controllers
 
             try
             {
+                string[] nome = user.Nome.ToLower().Split();
+                user.Senha = $"Hospital@{nome[0]}.{nome[nome.Length - 1]}";
+
                 await _context.Usuarios.AddAsync(user);
                 await _context.SaveChangesAsync();
+                Login login = new()
+                {
+                    Email = user.Email,
+                    Senha = user.Senha,
+                };
                 return CreatedAtAction(nameof(SelecionarUsuario),
-                    new { id = user.Id }, user);
+                    new { id = user.Id }, login);
             }
             catch (Exception ex)
             {
@@ -157,6 +167,62 @@ namespace DesafioFinal.Server.Controllers
                         Nome = user.Nome,
                         Email = user.Email
                     });
+                    _context.Usuarios.Remove(user);
+                    await _context.SaveChangesAsync();
+                    return NoContent();
+                }
+                catch (Exception ex)
+                {
+                    return BadRequest("Ocorreu um erro interno ao tentar efetuar o cadastro do usuário.\n" +
+                        "Erro:\n\t" + ex.Message);
+                }
+            }
+            return NotFound("Usuário não encontrado!");
+        }
+
+        /// <summary>
+        ///     Efetua login
+        /// </summary>
+        /// <param name="login">Login</param>
+        /// <returns>Usuário que condiz com o login</returns>
+        /// <response code="200">Logado!</response>
+        /// <response code="400">Erro ao efetuar o login!</response>
+        /// <response code="404">Erro de login!</response>
+        [AllowAnonymous]
+        [HttpPost("Login")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        public async Task<IActionResult> Logar([FromBody] Login login)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest("Ocorreu um erro ao efetuar o login!");
+
+            Usuario user = _context.Usuarios
+                .AsNoTracking()
+                .FirstOrDefault(
+                    user => user.Email == login.Email && user.Senha == login.Senha
+                );
+
+            if (user != null) 
+                return Ok(user);
+
+            return NotFound("Esse usuario não exite!");
+        }
+
+        [HttpPost("Logout/{id}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        public async Task<IActionResult> Logout([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest("Ocorreu um erro ao tentar efetuar o log-out do usuário.");
+
+            Usuario user = await _context.Usuarios
+                .AsNoTracking()
+                .FirstOrDefaultAsync(user => user.Id == id);
+
+            if (user != null)
+            {
+                try
+                {
                     _context.Usuarios.Remove(user);
                     await _context.SaveChangesAsync();
                     return NoContent();
